unit test for extension
